Allow setting product_type field for output product
This was previously set to the statistic name, which doesn't really work for
configurable statistics like `simple` or `external`. Now you have to supply
`product_type` field in the output product, failing that `product_type` will be
set to `!!NOTSET!!` string.

The correctness of the value in this field is only important when outputs are
destined to be indexed into the datacube, but not all uses of datacube-stats are
for generating products, so I think having a default is better than forcing
everyone to care about `product_type` field when creating configuration.