Handle `nodata` gracefully (#18)
Currently when a mask is applied, the input measurements
are converted into floats. The `nodata` values are
represented by `NaN`s in the float arrays and variable attributes
and stripped out. After calculating the statistic, we need to
cast the statistic to the intended type as reported by the
`measurements()` method, and convert `NaN`s back to
`nodata` values.

- In `Medoid` use `NaN` when `nodata` field is missing
- Add function `cast_back` to `datacube_stats.utils`
- Apply this function to output data in `load_process_save_chunk`