Publish successful build artifacts to S3
Can then be installed and upgraded using:

    pip install --extra-index-url https://packages.dea.gadevs.ga/ <packagename>

S3PyPi was a dead end

It didn't support our complex version numbers, and fixing it
would be a lot of work.

Here we can upload directly to S3, and a λ function keeps `index.html`
up to date. `pip` doesn't need anything else to be able to install
said packages.